Default `lights` of the `apply` service
The apply service is meant to apply the adaptive lighting parameters to a specific set of light(s). This set of lights need to be passed to the service, even though each Adaptive Lighting switch already has this list configured on itself. While having the flexibility to apply to some, it also might be useful to apply to all lights that the switch manages.

This patch makes the lights paramater optional and defaults it to the lights configured on the corresponding configuration entry of the switch being called.